+++ SitumWayfinding/Classes/Positioning/Internal/MarkerRenderer.swift
@@ -10,34 +10,39 @@ import GoogleMapsUtils
 class MarkerRenderer {
     var isClusteringEnabled: Bool { return markerClustering != nil }
     
-    private(set) var markers: Array<SitumMarker> = []
+    private(set) var markers: [SitumMarker] = []
     private var mapView: GMSMapView
-    private var buildingInfo: SITBuildingInfo
+    private var buildingManager: BuildingManager
     private var iconsStore: IconsStore
     private var showPoiNames: Bool
     private var markerClustering: GoogleMapsMarkerClustering? = nil
-    private var currentFloor: SITFloor?
+
+    private var currentFloor: SITFloor? = nil
+    private var selectedPoi: SITPOI? = nil
     
     init(
         mapView: GMSMapView,
-        buildingInfo: SITBuildingInfo,
+        buildingManager: BuildingManager,
         iconsStore: IconsStore,
         showPoiNames: Bool,
         isClusteringEnabled: Bool
     ) {
         self.mapView = mapView
-        self.buildingInfo = buildingInfo
+        self.buildingManager = buildingManager
         self.showPoiNames = showPoiNames
         self.iconsStore = iconsStore
         if isClusteringEnabled {
             markerClustering = GoogleMapsMarkerClustering(mapView: mapView)
         }
+        self.buildingManager.addDelegate(self)
     }
     
-    func displayPOIMarkers(forFloor floor: SITFloor) {
+    func displayPoiMarkers(forFloor floor: SITFloor) {
+        currentFloor = floor
         removeMarkers()
-        let poisInFloor = buildingInfo.indoorPois.filter { poi in poi.position().floorIdentifier == floor.identifier }
-        
+        var poisInFloor = buildingManager.filterPoisByCategories().filter(by: floor)
+        poisInFloor = preserveSelectedPoi(pois: poisInFloor, floor: floor)
+
         for poi in poisInFloor {
             let marker = SitumMarker(poi)
             markers.append(marker)
@@ -49,6 +54,22 @@ class MarkerRenderer {
                     self?.insertMarkerInGoogleMaps(marker: marker)
                 }
             }
+            selectMarkerIfIsSelectedPoi(marker: marker, poi: poi)
+        }
+    }
+
+    private func preserveSelectedPoi(pois: [SITPOI], floor: SITFloor) -> [SITPOI] {
+        guard let selectedPoi = selectedPoi else { return pois }
+        var innerPois = pois
+        if selectedPoi.belongs(to: floor) && !innerPois.contains(selectedPoi) {
+            innerPois.append(selectedPoi)
+        }
+        return innerPois
+    }
+
+    private func selectMarkerIfIsSelectedPoi(marker: SitumMarker, poi: SITPOI) {
+        if let selectedPoi = selectedPoi, selectedPoi.identifier == poi.identifier {
+            selectMarker(marker)
         }
     }
     
@@ -98,6 +119,7 @@ class MarkerRenderer {
             loadIcon(forMarker: marker, selected: true) { [weak self] marker in
                 self?.selectMarkerInGoogleMaps(marker: marker)
             }
+            selectedPoi = marker.poi
         } else if marker.isCustomMarker {
             selectMarkerInGoogleMaps(marker: marker)
         }
@@ -106,7 +128,7 @@ class MarkerRenderer {
     func isClusterGMSMarker(_ marker: GMSMarker) -> Bool {
         return isClusteringEnabled && marker.userData is GMUCluster
     }
-    
+
     func deselectMarker(_ marker: SitumMarker) {
         if marker.isPoiMarker {
             loadIcon(forMarker: marker, selected: false) { [weak self] marker in
@@ -121,6 +143,8 @@ class MarkerRenderer {
             removeMarkerFromGoogleMaps(marker: marker)
             markers = markers.filter { element in element != marker }
         }
+        selectedPoi = nil
+        removeMarkerIfPoiIsFiltered(marker)
     }
     
     private func selectMarkerInGoogleMaps(marker: SitumMarker) {
@@ -131,21 +155,28 @@ class MarkerRenderer {
     private func insertMarkerInGoogleMaps(marker: SitumMarker) {
         if marker.gmsMarker.map == nil {
             marker.setMapView(mapView: mapView)
-            marker.gmsMarker.zIndex = zIndices.poiMarker
+            marker.gmsMarker.zIndex = ZIndices.poiMarker
         }
     }
-    
+
     private func removeMarkerFromGoogleMaps(marker: SitumMarker) {
         if mapView.selectedMarker == marker.gmsMarker {
             deselectMarkerFromGoogleMaps(marker: marker)
         }
         marker.setMapView(mapView: nil)
     }
-    
+
     private func deselectMarkerFromGoogleMaps(marker: SitumMarker) {
         mapView.selectedMarker = nil
     }
-    
+
+    private func removeMarkerIfPoiIsFiltered(_ marker: SitumMarker) {
+        guard let poi = marker.poi, let currentFloor = currentFloor else { return }
+        if !buildingManager.hasCategoryIdInFilters(poi.categoryIdentifier) {
+            displayPoiMarkers(forFloor: currentFloor)
+        }
+    }
+
     /**
      Load the poi icon from the server for given marker
      - Parameters:
@@ -157,19 +188,21 @@ class MarkerRenderer {
         guard let poi = marker.poi else { return }
         // do not load images for markers that are not currently rendered (internal array)
         // also, ensure the marker is a reference to internal array of SitumMarker
-        guard let marker = searchMarker(byPOI: poi) else { return }
+        guard let marker = searchMarker(byPoi: poi) else { return }
         let showPoiNames = showPoiNames
         
         iconsStore.obtainIconFor(category: poi.category) { items in
-            guard let icon = selected ? items?[1] : items?[0] else {
+            guard var icon = selected ? items?[1] : items?[0] else {
                 Logger.logDebugMessage("Icon from server could not be retrieved")
                 return
             }
+            
+            icon = ImageUtils.scaleImageToSize(image: icon, newSize: CGSize(width: 45, height: 45))
     
             let color = UIColor(hex: "#5b5b5bff") ?? UIColor.gray
-            let title = poi.name.uppercased()
+            let title = poi.name
             if showPoiNames {
-                marker.gmsMarker.icon = icon.setTitle(title: title, size: 12.0, color: color, weight: .semibold)
+                marker.gmsMarker.icon = icon.setTitle(title: title, size: 16.0, color: color, weight: .bold)
             } else {
                 marker.gmsMarker.icon = icon
             }
@@ -181,10 +214,18 @@ class MarkerRenderer {
         return markers.first(where: { poiMarker in poiMarker.gmsMarker == gmsMarker })
     }
     
-    func searchMarker(byPOI searchedPoi: SITPOI) -> SitumMarker? {
+    func searchMarker(byPoi searchedPoi: SITPOI) -> SitumMarker? {
         return markers.first { marker in
             guard let poi = marker.poi else { return false }
             return poi.id == searchedPoi.id
+        }
+    }
+}
+
+extension MarkerRenderer: BuildingManagerDelegate {
+    func poiFiltersByCategoriesWereUpdated() {
+        if let floor = currentFloor {
+            displayPoiMarkers(forFloor: floor)
         }
     }
 }