+++ SitumWayfinding/Classes/Positioning/Navigation/InfoBarNavigationViewController.swift
@@ -48,9 +48,9 @@ class InfoBarNavigationViewController: UIViewController {
             self.setLoadingState()
             self.timeRemainingLabel.text = progress.currentIndication.humanReadableMessage()
         } else {
-            self.timeRemainingLabel.text = self.formatTime(time: progress.timeToGoal)
-            self.distanceRemainingLabel.text = self.formatDistance(distance: progress.timeToGoal)
-            self.estimatedTimeLabel.text = self.calculateEstimatedTime(timeToGoal: progress.timeToGoal)
+            self.timeRemainingLabel.text = self.formatTime(time: progress.improvedTimeToGoal)
+            self.distanceRemainingLabel.text = self.formatDistance(distance: progress.distanceToGoal)
+            self.estimatedTimeLabel.text = self.calculateEstimatedTime(timeToGoal: progress.improvedTimeToGoal)
         }
     }
     
@@ -78,20 +78,20 @@ class InfoBarNavigationViewController: UIViewController {
         if time.hours > 0 {
             let format = NSLocalizedString("navigation.remainingHours", bundle: SitumMapsLibrary.bundle,
                 comment: "Remaining time in hours")
-            return String(format: format, time.hours)
+            return String.localizedStringWithFormat(format, time.hours)
         } else if time.minutes > 0 {
             let format = NSLocalizedString("navigation.remainingMinutes", bundle: SitumMapsLibrary.bundle,
                 comment: "Remaining time in minutes")
-            return String(format: format, time.minutes)
+            return String.localizedStringWithFormat(format, time.minutes)
         } else {
             let format = NSLocalizedString("navigation.lessThanMinutes", bundle: SitumMapsLibrary.bundle,
                 comment: "Less than x minutes")
-            return String(format: format, 1)
+            return String.localizedStringWithFormat(format, 1)
         }
     }
     
     private func formatDistance(distance: Float) -> String {
-        return String(format: "%.1fm", distance)
+        return String.localizedStringWithFormat("%.1fm", distance)
     }
     
     private func secondsToHoursMinutesSeconds(_ seconds: Float) -> Time {
@@ -110,5 +110,14 @@ class InfoBarNavigationViewController: UIViewController {
         let formatter = DateFormatter()
         formatter.dateFormat = "HH:mm"
         return formatter.string(from: goalTime)
+    }
+}
+
+//TODO: This has to be deleted when this change is added to the sdk
+extension SITNavigationProgress {
+    var improvedTimeToGoal: Float {
+        get {
+            return timeToGoal/1.4
+        }
     }
 }