+++ SitumWayfinding/Classes/WayfindingLibrary/LibrarySettings.swift
@@ -41,8 +41,17 @@ import GoogleMaps
     /// Boolean that hides the back button in the navBar
     private (set) var showBackButton: Bool? = false
     
-    /// Boolean that controls if clustering of pois is enabled
+    /// Boolean that controls if the navigation indications are shown or hidden
     private (set) var enablePoisClustering: Bool = false
+    
+    /// Boolean to show or hide navigation indications
+    private (set) var showNavigationIndications: Bool = true
+    
+    /// Boolean to show or hide floors list
+    private (set) var floorsListVisible: Bool = true
+    
+    /// Boolean to show or hide positioning button
+    private (set) var positioningFabVisible: Bool = true
 
     // private(set) var orgDetails: OrganizationTheme?
     private override init() {
@@ -141,6 +150,27 @@ import GoogleMaps
             instance.enablePoisClustering = enablePoisClustering
             return self
         }
+        
+        /// Sets whether the navigation indications is visible
+        @discardableResult
+        @objc public func setShowNavigationIndications(showNavigationIndications: Bool) -> Builder {
+            instance.showNavigationIndications = showNavigationIndications
+            return self
+        }
+        
+        /// Sets whether the floors list is visible
+        @discardableResult
+        @objc public func setFloorsListVisible(floorsListVisible: Bool) -> Builder {
+            instance.floorsListVisible = floorsListVisible
+            return self
+        }
+        
+        /// Sets whether the positioning button is visible
+        @discardableResult
+        @objc public func setPositioningFabVisible(positioningFabVisible: Bool) -> Builder {
+            instance.positioningFabVisible = positioningFabVisible
+            return self
+        }
 
         /// Returns an instance of LibrarySettings
         @objc public func build() -> LibrarySettings {
@@ -185,6 +215,12 @@ import GoogleMaps
             builderCopy.setShowBackButton(showBackButton: settings.showBackButton ?? false)
 
             builderCopy.setEnablePoiClustering(enablePoisClustering: settings.enablePoisClustering)
+            
+            builderCopy.setShowNavigationIndications(showNavigationIndications: settings.showNavigationIndications)
+            
+            builderCopy.setFloorsListVisible(floorsListVisible: settings.floorsListVisible)
+            
+            builderCopy.setPositioningFabVisible(positioningFabVisible: settings.positioningFabVisible)
 
             return builderCopy
         }